backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
openai==1.3.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pdfplumber==0.10.3
python-dateutil==2.8.2
pyocrspace==0.1.5
